+++ practice/medium/candy/candy_test.go
@@ -17,7 +17,7 @@ import "testing"
 
 // TestCandiesFirstGivenCase implements the test given as first example on hackerrank.
 func TestCandiesFirstGivenCase(t *testing.T) {
-	n := int64(10)
+	n := 10
 	arr := []int64{2, 4, 2, 6, 1, 7, 8, 9, 2, 1}
 	expected := int64(19)
 
@@ -30,7 +30,7 @@ func TestCandiesFirstGivenCase(t *testing.T) {
 
 // TestCandiesSecondGivenCase implements the test given as second example on hackerrank.
 func TestCandiesSecondGivenCase(t *testing.T) {
-	n := int64(8)
+	n := 8
 	arr := []int64{2, 4, 3, 5, 2, 6, 4, 5}
 	expected := int64(12)
