+++ sort/easy/two/insertion.go
@@ -19,8 +19,8 @@ import (
 )
 
 // insertionSort print sorted values of the given array.
-func insertionSort(w io.Writer, n int64, arr []int64) {
-	for i := 1; i < int(n); i++ {
+func insertionSort(w io.Writer, n int, arr []int64) {
+	for i := 1; i < n; i++ {
 		temp := arr[i]
 		index := i - 1
