+++ structures/medium/waiter/waiter_test.go
@@ -21,7 +21,7 @@ import (
 // TestWaiterFirstGivenCase implements the test given as first example on hackerrank.
 func TestWaiterFirstGivenCase(t *testing.T) {
 	number := []int64{3, 4, 7, 6, 5}
-	q := int64(1)
+	q := 1
 
 	expected := []int64{4, 6, 3, 7, 5}
 
@@ -35,7 +35,7 @@ func TestWaiterFirstGivenCase(t *testing.T) {
 // TestWaiterSecondGivenCase implements the test given as second example on hackerrank.
 func TestWaiterSecondGivenCase(t *testing.T) {
 	number := []int64{3, 3, 4, 4, 9}
-	q := int64(2)
+	q := 2
 
 	expected := []int64{4, 4, 9, 3, 3}
